refactor biz logic, fix groups importer, fix conditions, create attributes as needed
