+++ packages/front-end/test/services/importing.test.ts
@@ -4,7 +4,7 @@ import {
   transformLDEnvironmentsToGBEnvironment,
   transformLDFeatureFlagToGBFeature,
   transformLDProjectsToGBProject,
-} from "@/services/importing";
+} from "@/services/importing/launchdarkly/launchdarkly-importing";
 
 describe("importing utils", () => {
   describe("transformLDEnvironmentsToGBEnvironment", () => {
